Add rhel6 to build_ami

+++ scripts/build_ami.sh
@@ -3,8 +3,15 @@
 #
 function make_link()
 {
-  if [ ! -e $1/x86_64-linux ]; then
-    ln -s x86_64-linux-opt $1/x86_64-linux
+  if [ -e $1/x86_64-linux-opt ]; then
+    if [ ! -e $1/x86_64-linux ]; then
+      ln -s x86_64-linux-opt $1/x86_64-linux
+    fi
+  fi
+  if [ -e $1/x86_64-rhel6-opt ]; then
+    if [ ! -e $1/x86_64-rhel6 ]; then
+      ln -s x86_64-rhel6-opt $1/x86_64-rhel6
+    fi
   fi
 }
 
@@ -12,6 +19,8 @@ make i386-linux-opt
 make i386-linux-dbg
 make x86_64-linux-opt
 make x86_64-linux-dbg
+make x86_64-rhel6-opt
+make x86_64-rhel6-dbg
 
 make_link build/ami/lib
